requirements.txt
Flask
pandas
matplotlib
pytest
python-dotenv
sqlalchemy
celery
redis
loguru